requirements.txt
pandas>=2.2,<3
numpy>=1.26,<2
pyyaml>=6
matplotlib>=3.8
scipy>=1.11
requests>=2.32
python-dotenv>=1.0
typer>=0.12,<1
pydantic>=2.8,<3
pyarrow>=17,<18
